refactor(FON-468): dissolve secretariat-general components into domain features
