+++ apps/client/src/components/shared/ui/document-preview/DocumentEditor.css
