+++ apps/client/src/components/shared/ui/document-preview/DocumentPreview.tsx
