+++ apps/client/src/features/members/components/DetailsMemberHeader.tsx
@@ -3,8 +3,8 @@ import Button from '@codegouvfr/react-dsfr/Button';
 import React from 'react';
 import { useNavigate } from 'react-router';
 
-import { ROUTE_PATHS } from '../../../../utils/route-path.utils';
-import { capitalize } from '../../../../utils/string.utils';
+import { ROUTE_PATHS } from '@/utils/route-path.utils';
+import { capitalize } from '@/utils/string.utils';
 
 export function DetailsMemberHeader(props: { member: { firstName: string; lastName: string } }) {
   const navigate = useNavigate();
