+++ apps/client/src/features/sessions/components/ManageSession.tsx
