+++ apps/client/src/features/sessions/components/SessionStatusBadge.tsx
