+++ apps/client/src/features/transparence/components/DocActionDelete.tsx
