+++ apps/client/src/features/transparence/components/DocActionDetails.tsx
