+++ apps/client/src/features/transparence/components/DocGenerationAction.tsx
