+++ apps/client/src/features/transparence/components/NouvelleTransparence.tsx
@@ -12,12 +12,12 @@ import { FormattedMessage } from 'react-intl';
 import { useNavigate } from 'react-router';
 import { z } from 'zod';
 
-import { ROUTE_PATHS } from '../../../../utils/route-path.utils';
-import { getSgBreadCrumb } from '../../../../utils/sg-breadcrumb.utils';
-import { Breadcrumb } from '../../../shared/Breadcrumb';
-import { PageContentLayout } from '../../../shared/PageContentLayout';
+import { Breadcrumb } from '@/components/shared/Breadcrumb';
 import { Mandatory } from '@/components/shared/Mandatory';
+import { PageContentLayout } from '@/components/shared/PageContentLayout';
 import { FormationEnum, FormationEnumLabel } from '@/types/enums.types';
+import { ROUTE_PATHS } from '@/utils/route-path.utils';
+import { getSgBreadCrumb } from '@/utils/sg-breadcrumb.utils';
 import { capitalize } from '@/utils/string.utils';
 import { useCreateNominationSessionFromLodamMutation } from '@queries/nomination-sessions.queries';
