+++ apps/client/src/features/transparence/components/TableauDeBordResumeDetails.tsx
